requirements.txt
liac_arff==2.5.0
pandas==2.2.3
streamlit==1.45.1
